fix(constants): move named base component exports to root constants file
Turns out importing the named exports from *.vue files does not go over too well.
